Type: refactor Moved SGD & Adam to a single file & tested
